+++ client/src/pages/home/Home.tsx
@@ -6,11 +6,11 @@ import axios from "axios";
 import { Spinner } from "@chakra-ui/react";
 import { ICity } from "../../assets/interfaces/City";
 import NearbyCityCard from "./components/NearbyCityCard";
+import Layout from "../../components/Layout";
 
 const Home = () => {
   const [open, setOpen] = useState(false);
   const [nearbyCities, setNearbyCities] = useState<ICity[]>([]);
-  const [user, setUser] = useState(null);
 
   const toggleSidebar = () => {
     setOpen(!open);
@@ -44,40 +44,12 @@ const Home = () => {
     }
   };
 
-  const getUserInfo = async () => {
-    const token = localStorage.getItem("token");
-
-    try {
-      const response = await axios.get(
-        "http://localhost:3001/login/verifyToken",
-        {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        }
-      );
-      if (response.data.user) {
-        console.log(response.data);
-        setUser(response.data.user);
-      }
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
   useEffect(() => {
     fetchNearbyCities();
-    getUserInfo();
   }, []);
 
   return (
-    <div className="w-full flex flex-col h-screen items-center">
-      <div
-        className={`fixed w-full z-10 h-full bg-black bg-opacity-50 sm:bg-opacity-0 duration-100 ${
-          open ? "block sm:hidden" : "hidden"
-        }`}
-      />
-      <Navbar toggleSidebar={toggleSidebar} />
+    <Layout>
       <CurrCity getCoordinates={getCoordinates} />
       {nearbyCities.length > 0 ? (
         <div className="w-full lg:w-2/3 h-1/4">
@@ -100,9 +72,7 @@ const Home = () => {
           </h1>
         </div>
       )}
-
-      <Sidebar open={open} toggleSidebar={toggleSidebar} user={user} />
-    </div>
+    </Layout>
   );
 };
