+++ client/src/pages/planYourTrip/PlanPage.tsx
@@ -1,5 +1,6 @@
 import { BiSolidPlaneAlt, BiSolidHotel } from "react-icons/bi";
 import PlanButton from "./components/PlanButton";
+import Layout from "../../components/Layout";
 
 const options = [
   {
@@ -16,8 +17,8 @@ const options = [
 
 const PlanPage = () => {
   return (
-    <div className="w-full h-screen flex flex-col justify-center p-6">
-      <div className="w-full h-20 flex justify-center text-3xl md:text-4xl font-bold">
+    <Layout>
+      <div className="w-full h-20 flex justify-center text-3xl md:text-4xl font-bold mt-8">
         Plan the perfect trip!
       </div>
       <div className="w-full flex justify-evenly items-center">
@@ -29,7 +30,7 @@ const PlanPage = () => {
           />
         ))}
       </div>
-    </div>
+    </Layout>
   );
 };
